+++ src/main/java/com/picimako/obsidian/translation/OriginalLocalizationValuesCache.java
@@ -2,12 +2,15 @@ package com.picimako.obsidian.translation;
 
 import com.intellij.openapi.components.Service;
 import com.intellij.openapi.project.Project;
+import lombok.Getter;
+import lombok.Setter;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Project service for storing the original, English localization values each mapped to its property path
+ * Project service for storing the original, English localization values, each mapped to its property path
  * inside the translation files.
  * <p>
  * This is put into place, so that the English JSON file's PSI don't have to be re-read and traversed every time
@@ -19,15 +22,14 @@ public final class OriginalLocalizationValuesCache {
     /**
      * JSON property path -> original English localization value.
      */
+    @Getter
     private final Map<String, String> originalValues = new HashMap<>();
+    @Getter
+    @Setter
     private boolean isProjectObsidianTranslations;
 
     @SuppressWarnings("unused")
     public OriginalLocalizationValuesCache(Project project) {
-    }
-
-    public Map<String, String> getOriginalValues() {
-        return originalValues;
     }
 
     /**
@@ -37,19 +39,16 @@ public final class OriginalLocalizationValuesCache {
         originalValues.put(path, value);
     }
 
-    public boolean isProjectObsidianTranslations() {
-        return isProjectObsidianTranslations;
-    }
-
-    public void setProjectObsidianTranslations(boolean projectObsidianTranslations) {
-        isProjectObsidianTranslations = projectObsidianTranslations;
-    }
-
     public static OriginalLocalizationValuesCache getInstance(Project project) {
         return project.getService(OriginalLocalizationValuesCache.class);
     }
 
     public static boolean isProjectObsidianTranslations(Project project) {
         return OriginalLocalizationValuesCache.getInstance(project).isProjectObsidianTranslations();
+    }
+
+    @Nullable
+    public static String getOriginalValueAt(String propertyPath, Project project) {
+        return OriginalLocalizationValuesCache.getInstance(project).getOriginalValues().get(propertyPath);
     }
 }