+++ src/main/java/com/picimako/obsidian/translation/TranslationFileUtil.java
@@ -1,8 +1,5 @@
 package com.picimako.obsidian.translation;
 
-import static com.intellij.openapi.application.ReadAction.compute;
-import static com.intellij.psi.util.PsiTreeUtil.collectParents;
-
 import com.intellij.json.psi.JsonFile;
 import com.intellij.json.psi.JsonObject;
 import com.intellij.json.psi.JsonProperty;
@@ -11,7 +8,10 @@ import com.intellij.psi.PsiFile;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
+
+import static com.intellij.openapi.application.ReadAction.compute;
+import static com.intellij.psi.util.PsiTreeUtil.collectParents;
+import static java.util.stream.Collectors.toList;
 
 /**
  * Various utilities for JSON translation files.
@@ -21,16 +21,26 @@ public final class TranslationFileUtil {
     public static final String ESLINTRC_JSON = ".eslintrc.json";
 
     /**
+     * Gets the path of the property at {@code startElement}, each property name delimited with a dot symbol.
+     *
+     * @param startElement a sub-element of a {@link JsonProperty} whose path is retrieved
+     * @param endElement used for the stop condition as the last element to stop the parent property collection at
+     */
+    public static String getPropertyPath(PsiElement startElement, PsiElement endElement) {
+        return String.join(".", getPropertyPathElements(startElement, endElement));
+    }
+
+    /**
      * Gets the path of the property as list of property names leading to it.
      *
      * @param startElement the element to get the path of
      * @param endElement   used for the stop condition as the last element to stop the parent property collection at
      */
-    public static List<String> getPropertyPath(PsiElement startElement, PsiElement endElement) {
+    public static List<String> getPropertyPathElements(PsiElement startElement, PsiElement endElement) {
         var propertyPath = compute(() -> collectParents(startElement, JsonProperty.class, true, e -> e.isEquivalentTo(endElement)))
             .stream()
             .map(JsonProperty::getName)
-            .collect(Collectors.toList());
+            .collect(toList());
         Collections.reverse(propertyPath);
 
         return propertyPath;
@@ -43,7 +53,7 @@ public final class TranslationFileUtil {
      * @param topLevelObject the top level object in a translation file
      */
     public static JsonProperty findPropertyByPath(List<String> path, JsonObject topLevelObject) {
-        var property = compute(() -> topLevelObject.findProperty(path.get(0)));
+        var property = compute(() -> topLevelObject.findProperty(path.getFirst()));
 
         for (int i = 1; i < path.size(); i++) {
             final int index = i;